requirements.txt
pandas
mysql-connector-python
streamlit
plotly
gitpython
